Move grpah implementations to own packages
