+++ src/main/java/de/hpi/msd/salsa/graph/adjacencyList/DistributedKeyValueGraph.java
@@ -1,6 +1,7 @@
-package de.hpi.msd.salsa.graph;
+package de.hpi.msd.salsa.graph.adjacencyList;
 
 import de.hpi.msd.salsa.EdgeToAdjacencyApp;
+import de.hpi.msd.salsa.graph.KeyValueGraph;
 import de.hpi.msd.salsa.serde.avro.AdjacencyList;
 import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.KafkaStreams;
